Refactor code to categorize images better (#48)
