+++ src/Images/Landing/icon-white.png
