+++ src/Pages/Home/home.css
@@ -1,0 +1,4 @@
+.home-page {
+    color: black;
+    
+}
